// packages/contract/packages/contract/src/errors.d.ts
export declare const ERRORS: {
    CONTRACT: {
        INVALID_METHOD: {
            message: string;
        };
        TX_ERROR: {
            message: string;
        };
        QUERY_ERROR: {
            message: string;
        };
        INVALID_ADDRESS: {
            message: string;
        };
        INVALID_STORAGE_NAME: {
            message: string;
        };
        CAPTCHA_SOLUTION_COMMITMENT_DOES_NOT_EXIST: {
            message: string;
        };
        DAPP_NOT_ACTIVE: {
            message: string;
        };
        CONTRACT_UNDEFINED: {
            message: string;
        };
        SIGNER_UNDEFINED: {
            message: string;
        };
    };
};